@weavejester 2023-10-03
Thanks for the PR. Those are some very helpful optimizations. To answer your questions:

> Why are we sorting the attributes and the style attribute? Is this just for test output?

It's sorted to produce a consistent output, so a particular input will always produce the same output. If this weren't the case, the attribute order could change, as map keys are unordered.

> Why do we escaping HTML text/attributes by default? Isn't this a complex operation?

For safety. The consequences of client data being unescaped is significant enough that it's worth erring on the side of security over speed in this case.