@Panthevm 2023-10-03
Thanks for the answers! @weavejester

> For safety. The consequences of client data being unescaped is significant enough that it's worth erring on the side of security over speed in this case.

There is an option called **_escape-strings?_**, but it isn't considered when generating HTML attribute values. Was this intentional? 

I think there should be a call to the hiccup.compiler/escape-html function there

https://github.com/weavejester/hiccup/blob/5318b5ea73bffa76f03a4b3add184f50f9b9dd8a/src/hiccup/compiler.clj#L40-L41