@weavejester 2023-10-03
> There is an option called ***escape-strings?***, but it isn't considered when generating HTML attribute values. Was this intentional?

Yes. This is because there's no reason not to escape attribute values, whereas strings in tags may potentially contain other elements.