@Panthevm 2023-10-03
> Yes. This is because there's no reason not to escape attribute values, whereas strings in tags may potentially contain other elements.

Escape strings are time-consuming. I'd like the ability to disable them, specifying escape locations only where necessary.

It's possible to turn them off everywhere, except in attribute values, which seems odd.
```
 (def value "SSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSS")
 (criterium.core/quick-bench (hiccup.util/escape-html value))

;;  Variance from outliers : 13.8889 % Variance is moderately inflated by outliers
;; Evaluation count : 5124636 in 6 samples of 854106 calls.
;;              Execution time mean : 110.759692 ns
;;     Execution time std-deviation : 1.323253 ns
;;    Execution time lower quantile : 109.373272 ns ( 2.5%)
;;    Execution time upper quantile : 112.215336 ns (97.5%)
;;                    Overhead used : 7.985989 ns
```