@weavejester 2023-10-03
We could add a new option, something like `disable-all-escaping?`, that would also disable escaping in attribute values. We can't change the existing behavior because existing users may be relying on it.