@Panthevm 2023-10-03
I found another unnecessary seq call in hiccup.compiler/normalize-element

Profiling report:
<img width="1138" alt="image" src="https://github.com/weavejester/hiccup/assets/43318093/9eb8727c-2e73-4897-ade8-4a00ffe11161">

1247716c3b72eef282e847b901baf228715df597