@Panthevm 2023-10-03
> It's sorted to produce a consistent output, so a particular input will always produce the same output. If this weren't the case, the attribute order could change, as map keys are unordered.

What about this optimization? Sort only when order is not guaranteed

``` patch
diff --git a/src/hiccup/compiler.clj b/src/hiccup/compiler.clj
index 43adcfc..f265bbe 100644
--- a/src/hiccup/compiler.clj
+++ b/src/hiccup/compiler.clj
@@ -43,7 +43,9 @@
        (.append sb ":")
        (.append sb (util/to-str v))
        (.append sb ";"))
-     (sort-by #(name (key %)) value))
+     (if (instance? clojure.lang.PersistentArrayMap value)
+       value
+       (sort-by #(name (key %)) value)))
     (.toString sb)))

 (defn- render-attr-value [value]
@@ -77,7 +79,9 @@
     ""
     (let [^StringBuilder sb (StringBuilder.)]
       (iterate! #(.append sb (render-attribute %))
-                (sort-by #(name (key %)) attrs))
+                (if (instance? clojure.lang.PersistentArrayMap attrs)
+                  attrs
+                  (sort-by #(name (key %)) attrs)))
       (.toString sb))))

 (def ^{:doc "A list of elements that must be rendered without a closing tag."
```
#### Benchmarks

<details>
  <summary>Before</summary>
  
``` clj
(def hiccup
  [:html
   [:head
    [:title "Title"]]
   [:body
    [:span "foo"]
    [:span#id "foo"]
    [:span#id.class1.class2 "foo"]
    [:span#id.class1.class2 {:id "id"} "foo"]
    [:span#id.class1.class2 {:class "class3"} "foo"]
    [:span#id.class1.class2 {:styles {:a 1 :b 2}} "foo"]
    (for [i [1 2 3 4]]
      [:span i])]])

(criterium.core/bench
  (hiccup2.core/html hiccup))
  
;; Evaluation count : 3157980 in 60 samples of 52633 calls.
;;              Execution time mean : 18.995492 µs
;;     Execution time std-deviation : 434.056573 ns
;;    Execution time lower quantile : 18.788194 µs ( 2.5%)
;;    Execution time upper quantile : 19.519848 µs (97.5%)
;;                    Overhead used : 8.242091 ns

;; Found 4 outliers in 60 samples (6.6667 %)
;; 	low-severe	 1 (1.6667 %)
;; 	low-mild	 3 (5.0000 %)
;;  Variance from outliers : 10.9878 % Variance is moderately inflated by outliers

```
</details>

<details>
  <summary>After</summary>
  
``` clj
(def hiccup
  [:html
   [:head
    [:title "Title"]]
   [:body
    [:span "foo"]
    [:span#id "foo"]
    [:span#id.class1.class2 "foo"]
    [:span#id.class1.class2 {:id "id"} "foo"]
    [:span#id.class1.class2 {:class "class3"} "foo"]
    [:span#id.class1.class2 {:styles {:a 1 :b 2}} "foo"]
    (for [i [1 2 3 4]]
      [:span i])]])

(criterium.core/bench
  (hiccup2.core/html hiccup))
  
;; Evaluation count : 3908700 in 60 samples of 65145 calls.
;;              Execution time mean : 15.321174 µs
;;     Execution time std-deviation : 116.528715 ns
;;    Execution time lower quantile : 15.154524 µs ( 2.5%)
;;    Execution time upper quantile : 15.545923 µs (97.5%)
;;                    Overhead used : 8.242091 ns

;; Found 2 outliers in 60 samples (3.3333 %)
;; 	low-severe	 1 (1.6667 %)
;; 	low-mild	 1 (1.6667 %)
;;  Variance from outliers : 1.6389 % Variance is slightly inflated by outliers

```
</details>