@Panthevm 2023-10-03
At the moment, the weakest point is accessing the dynamic var. Ideally, we should get rid of it and pass the original opts hash-map instead.

![image](https://github.com/weavejester/hiccup/assets/43318093/ff5ca121-66a2-42f9-a81d-84f7b330edce)