@weavejester 2023-10-03
The order is never guaranteed for a map. `{:a 1, :b 2}` and `{:b 2, :a 1}` are array maps with different key orderings internally, but are considered equal, i.e. `(= {:a 1, :b 2} {:b 2, :a 1})`.

It's important to ensure that `(= (html x) (html y))` if `(= x y)`.