@weavejester 2023-10-03
Maybe; that would be another PR though. I'm cautious of any option that makes functions less secure or predictable, even if it's off by default. Checking that option may also incur some small performance hit.