@Panthevm 2023-10-04
You can still use new functions when compile-form. This will give a performance gain. But this is already micro-optimization

``` clj
(def value "hello")

(criterium.core/bench
  (hiccup2.core/html
    [:html
     [:head
      [:title "Title"]]
     [:body
      [:span "foo"]
      [:span#id "foo"]
      value
      [:span#id.class1.class2 "foo"]
      [:span#id.class1.class2 {:id "id"} "foo"]
      [:span#id.class1.class2 {:class "class3"} "foo"]
      [:span#id.class1.class2 {:styles {:a 1 :b 2}} "foo"]]]))

;;;; Before
;;
;; Evaluation count : 86879760 in 60 samples of 1447996 calls.
;;              Execution time mean : 683.640477 ns
;;     Execution time std-deviation : 11.681363 ns
;;    Execution time lower quantile : 674.248617 ns ( 2.5%)
;;    Execution time upper quantile : 700.726863 ns (97.5%)
;;                    Overhead used : 8.340812 ns

;; Found 3 outliers in 60 samples (5.0000 %)
;; 	low-severe	 2 (3.3333 %)
;; 	low-mild	 1 (1.6667 %)
;;  Variance from outliers : 6.2702 % Variance is slightly inflated by outliers


;;;; After patch
;;
;; Evaluation count : 185077020 in 60 samples of 3084617 calls.
;;              Execution time mean : 312.203067 ns
;;     Execution time std-deviation : 7.959083 ns
;;    Execution time lower quantile : 306.728104 ns ( 2.5%)
;;    Execution time upper quantile : 325.430932 ns (97.5%)
;;                    Overhead used : 8.340812 ns

;; Found 5 outliers in 60 samples (8.3333 %)
;; 	low-severe	 3 (5.0000 %)
;; 	low-mild	 2 (3.3333 %)
;;  Variance from outliers : 12.6210 % Variance is moderately inflated by outliers

```

<details>
  <summary>Patch</summary>
  
``` patch
diff --git a/src/hiccup/compiler.clj b/src/hiccup/compiler.clj
index d8bf187..1367209 100644
--- a/src/hiccup/compiler.clj
+++ b/src/hiccup/compiler.clj
@@ -23,13 +23,13 @@
 (defn- end-tag []
   (if (xml-mode?) " />" ">"))
 
-(defn- iterate! [callback coll]
+(defn iterate! [callback coll]
   (when coll
     (let [^Iterator iterator (.iterator ^Iterable coll)]
       (while (.hasNext iterator)
         (callback (.next iterator))))))
 
-(defmacro ^:private build-string [& strs]
+(defmacro build-string [& strs]
   (let [w (gensym)]
     `(let [~w (StringBuilder.)]
        ~@(map (fn [arg] `(.append ~w ~arg)) strs)
@@ -222,7 +222,12 @@
 
 (defmethod compile-form "for"
   [[_ bindings body]]
-  `(apply str (for ~bindings ~(compile-html body))))
+  `(let [^StringBuilder sb# (StringBuilder.)]
+     (iterate! #(.append sb# %)
+               (for ~bindings ~(compile-html body)))
+     (.toString sb#))
+  ;;`(apply str (for ~bindings ~(compile-html body)))
+  )
 
 (defmethod compile-form "if"
   [[_ condition & body]]
@@ -301,10 +306,10 @@
   [[tag attrs & content]]
   (let [[tag attrs _] (normalize-element-form [tag attrs])]
     (if (container-tag? tag content)
-      `(str ~(str "<" tag) ~(compile-attr-map attrs) ">"
-            ~@(compile-seq content)
-            ~(str "</" tag ">"))
-      `(str "<" ~tag ~(compile-attr-map attrs) ~(end-tag)))))
+      `(build-string ~(build-string "<" tag) ~(compile-attr-map attrs) ">"
+                     ~@(compile-seq content)
+                     ~(build-string "</" tag ">"))
+      `(build-string "<" ~tag ~(compile-attr-map attrs) ~(end-tag)))))
 
 (defmethod compile-element ::literal-tag-and-no-attributes
   [[tag & content]]
@@ -317,18 +322,18 @@
     `(let [~attrs-sym ~attrs]
        (if (map? ~attrs-sym)
          ~(if (container-tag? tag content)
-            `(str ~(str "<" tag)
-                  (render-attr-map (merge ~tag-attrs ~attrs-sym)) ">"
-                  ~@(compile-seq content)
-                  ~(str "</" tag ">"))
-            `(str ~(str "<" tag)
-                  (render-attr-map (merge ~tag-attrs ~attrs-sym))
-                  ~(end-tag)))
+            `(build-string ~(build-string "<" tag)
+                           (render-attr-map (merge ~tag-attrs ~attrs-sym)) ">"
+                           ~@(compile-seq content)
+                           ~(build-string "</" tag ">"))
+            `(build-string ~(build-string "<" tag)
+                           (render-attr-map (merge ~tag-attrs ~attrs-sym))
+                           ~(end-tag)))
          ~(if (container-tag? tag attrs)
-            `(str ~(str "<" tag (render-attr-map tag-attrs) ">")
-                  ~@(compile-seq (cons attrs-sym content))
-                  ~(str "</" tag ">"))
-            (str "<" tag (render-attr-map tag-attrs) (end-tag)))))))
+            `(build-string ~(build-string "<" tag (render-attr-map tag-attrs) ">")
+                           ~@(compile-seq (cons attrs-sym content))
+                           ~(build-string "</" tag ">"))
+            (build-string "<" tag (render-attr-map tag-attrs) (end-tag)))))))
 
 (defmethod compile-element :default
   [element]
@@ -361,7 +366,9 @@
     (cons
      (first expr)
      (mapcat
-      #(if (and (seq? %) (symbol? (first %)) (= (first %) (first expr) `str))
+      #(if (and (seq? %) (symbol? (first %))
+                (or (= (first %) (first expr) `str)
+                    (= (first %) (first expr) `build-string)))
          (rest (collapse-strs %))
          (list (collapse-strs %)))
       (rest expr)))
@@ -370,7 +377,7 @@
 (defn compile-html
   "Pre-compile data structures into HTML where possible."
   [& content]
-  (collapse-strs `(str ~@(compile-seq content))))
+  (collapse-strs `(build-string ~@(compile-seq content))))
 
 (defn- binding* [var val func]
   (push-thread-bindings {var val})

```
  
</details>